feat: create new Component-ExtendNavBar
